test: add new page for both owncloud and opencloud
