+++ tests/e2e/stepDefinitions/presentationViewerContext.js
@@ -1,16 +1,15 @@
 const { Given, When, Then } = require('@cucumber/cucumber')
 const { expect } = require('@playwright/test')
 
-const config = require('../config')
-
 const { getUserCredentials } = require('../utils/userHelper')
 const { uploadFile } = require('../utils/fileHelper')
 const PresentationViewer = require('../pageObjects/PresentationViewerPage')
 const Ocis = require('../pageObjects/OcisPage')
-const Opencloud = require('../pageObjects/OpencloudPage')
+const Files = require('../pageObjects/FilesPage')
 
 const presentationViewer = new PresentationViewer()
 const ocis = new Ocis()
+const files = new Files()
 
 Given(
   'user {string} has uploaded the markdown file {string} using API',
@@ -29,12 +28,7 @@ Given('user {string} has logged in', async function (user) {
 When(
   'user {string} previews a markdown file {string} in presentation viewer',
   async function (user, fileName) {
-    if (config.targetApp === 'opencloud') {
-      const opencloud = new Opencloud()
-      await opencloud.openMDFileInPresentationViewer()
-    } else {
-      await ocis.openMDFileInPresentationViewer()
-    }
+    await files.openMDFileInPresentationViewer()
   }
 )
